893 | Fix "Encounter - Observation" data model in couch-mrs to improve performance
https://trello.com/c/TmGk7bFW/893-fix-encounter-observation-data-model-in-couch-mrs-to-improve-performance

The Encounter/Observation data model has been changed from an application managed relationship across
two databases in Couch to instead storing observations directly as part of the Encounter document. This
was done to reduce the number of HTTP requests to Couch when reconstructing the full object graph
of encounters with many observations.

Change-Id: I9ed7742b1e306c080c11726a73e95439b3c784f0